fix: amend bandsharing adjustment procedure
feat: get_PNLT bandsharing procedure was omitted if the Cmax(kM) index was near the start or end of the spectral time-series. This amendment averages the Cavg over the available Cmax values within the -2, +2 interval of kM using a logical mask. This seems a more reasonable approach than omitting the adjustment entirely if some indices are negative or outside the length of the series.